refactor: isolate pullRequestsSearch into separate hook file
